Move "derive_address" function to _crypto module
- To avoid circular import
- `_crypto` is the only usage
